Extract I/O from core.py
